refactor(remote): 호스트 후보 로직 중복 제거 + Rust 모듈 분리
PR #417 리뷰 후속 정리(동작 불변):
- 호스트 후보 fetch/guard/폴백/buildRemoteHostOptions 를 공용 훅
  useRemoteHostOptions 로 통합. 모달·SettingsView 중복 제거.
- Rust 호스트 후보 기능을 commands/remote_hosts.rs 로 분리(misc.rs 비대
  해소, AGENTS §14 500줄 규칙). mod 등록+re-export 로 handler 경로 유지.
- 모달 lastHost state→useRef 마운트 1회 읽기로 과잉 파생 제거.
