+++ src-tauri/src/commands/mod.rs
@@ -2,10 +2,12 @@ mod claude_session;
 mod file_ops;
 mod ipc_dispatch;
 mod misc;
+mod remote_hosts;
 mod terminal;
 
 pub use claude_session::*;
 pub use file_ops::*;
 pub use ipc_dispatch::*;
 pub use misc::*;
+pub use remote_hosts::*;
 pub use terminal::*;
